Added cancelation method
- Added primitive logging
- Added cancelation method to OperationManager
- Fixed int conversion warning

+++ AFNetworking/AFURLRequestSerialization.m
@@ -145,7 +145,7 @@ NSArray * AFQueryStringPairsFromKeyAndValue(NSString *key, id value) {
         NSArray *array = value;
 		NSUInteger index = 0;
         for (id nestedValue in array) {
-            [mutableQueryStringComponents addObjectsFromArray:AFQueryStringPairsFromKeyAndValue([NSString stringWithFormat:@"%@[%lu]", key, (unsigned long)index], nestedValue)];
+            [mutableQueryStringComponents addObjectsFromArray:AFQueryStringPairsFromKeyAndValue([NSString stringWithFormat:@"%@[%u]", key, index], nestedValue)];
 			++index;
         }
     } else if ([value isKindOfClass:[NSSet class]]) {
